updating matrix gen

+++ 3.generate-matrices.py
@@ -1,6 +1,10 @@
 #!/usr/bin/env python
 
 import pickle
+from glob import glob
+
+# module load python/3.6.1
+# ml py-ipython/6.1.0_py36
 
 # First, let's read in both years data and get a set of unique container uris
 
@@ -15,22 +19,101 @@ print(len(containers_2019))
 
 containers = set()
 
-containers.index(container)
-for container in containers_2017 + containers_2019:
-    containers.add(container)
+if os.path.exists('dockerfiles-2017+2019.pkl'):
+    containers = pickle.load(open('dockerfiles-2017+2019.pkl', 'rb'))
+
+else:
+    for container in containers_2017 + containers_2019:
+        containers.add(container)
+
+    pickle.dump(containers, open('dockerfiles-2017+2019.pkl','wb'))
 
 print(len(containers))
-
 # 505234
-pickle.dump(containers, open('dockerfiles-2017+2019.pkl','wb'))
-
 
 ################################################################################
-# Trees!
+# Pip Trees!
 ################################################################################
 
 # For each, build an apt and pip tree
-from containertree import ContainerPipTree, ContainerAptTree
+from containertree import ContainerPipTree
+
+pip = ContainerPipTree()
+
+seen = set()
+
+# Parse output files from container-diff
+diffs = glob('/scratch/users/vsochat/WORK/dockerfile-packages/container-diff/*.json')
+pickle.dump(diffs, open('diffs.pkl','wb'))
+
+# len(diffs)
+# 151427
+
+seen = []
+for diff in diffs:
+    tag = os.path.basename(diff).replace('-','/', 1).replace('.json','')
+    if tag in seen:
+        continue
+    seen.append(tag)
+    try:
+        pip.update(diff, tag=tag)
+    except TypeError:
+        pass
+
+
+count = 1000
+for container in containers:
+
+    if container in pip.root.tags:
+        continue
+
+    if count % 1000 == 0:
+        pickle.dump(pip, open('pip-tree.pkl', 'wb'))
+
+    if container in seen:
+        continue
+
+    print('Adding %s' % container)
+    try:
+        pip.update(container, tag=container)
+    except TypeError:
+        pass
+
+    seen.add(container)
+    count +=1
+
+pickle.dump(pip, open('pip-tree.pkl', 'wb'))
+pickle.dump(seen, open('seen-containers.pkl', 'wb'))
+
+# Stopped at count 53854
+# container
+# 'booyaabes/kali-linux-full'
+# see seen.pkl to pick up, we are good to start with a set of this size:
+# len(seen) -> 19,501
+# len(apt.root.tags) -> 6025
+# len(pip.root.tags) -> 3685
+
+# Generate matrices for both
+pip_vectors = pip.export_vectors()
+# pip_vectors.shape
+# (3685, 1917)
+pip_vectors = pip_vectors.fillna(0)
+pip_vectors.to_csv('pip-vectors.csv')
+pip_vectors.to_pickle('pip-vectors.pkl')
+
+apt_vectors = apt.export_vectors()
+# apt_vectors.shape
+# (6025, 17194)
+apt_vectors = apt_vectors.fillna(0)
+apt_vectors.to_csv('apt-vectors.csv')
+apt_vectors.to_pickle('apt-vectors.pkl')
+
+################################################################################
+# Apt Trees!
+################################################################################
+
+# For each, build an apt and pip tree
+from containertree import ContainerAptTree
 
 apt = ContainerAptTree()
 pip = ContainerPipTree()
